use sql_par_field_list and review sql_insert

+++ src/test/resources/db/log/change_insert.sql
@@ -1,4 +1,4 @@
-PREPARE change_insert (bigint, bigint, bigint, text, bigint) AS
+PREPARE change_insert (bigint, smallint, smallint, text, bigint) AS
     INSERT INTO changes
                 (user_id, change_action_id, change_field_id, new_value, row_id)
          VALUES ($1, $2, $3, $4, $5)
