+++ src/test/resources/db/ref/source_insert_0111110000.sql
@@ -1,4 +1,4 @@
-PREPARE source_insert_0111110000 (bigint, text, text, bigint, text) AS
+PREPARE source_insert_0111110000 (bigint, text, text, smallint, text) AS
     INSERT INTO sources (user_id, source_name, description, source_type_id, url)
          VALUES         ($1, $2, $3, $4, $5)
       RETURNING source_id;
